requirements.txt
mariadb==1.1.5.post3
packaging==22.0
python-decouple==3.6

Flask~=2.2.2
requests~=2.28.1
connexion~=2.14.1